Fixed a warning on Linux.
git-svn-id: svn://edison.comlab.ox.ac.uk/res07/stug/millipede@379 2d08482b-68a0-452d-bcbc-5be95be64364

+++ trunk/source/common/partitionforests/images/VolumeIPF.h
@@ -68,7 +68,7 @@ public:
 	{
 		for(int i=0; i<3; ++i)
 		{
-			if(position[i] < 0 || position[i] >= m_volumeSize[i])
+			if(position[i] < 0 || position[i] >= static_cast<long>(m_volumeSize[i]))
 			{
 				return -1;
 			}
